[cleanup] moves CodeGhost.js to containers
